+++ packages/utils/src/graphql/entities.ts
@@ -20,10 +20,12 @@ import {
   isUnionType,
   ValueNode,
   BooleanValueNode,
+  ListTypeNode,
+  TypeNode,
 } from 'graphql';
 import {findDuplicateStringArray} from '../array';
 import {Logger} from '../logger';
-import {getTypeByScalarName} from '../types';
+import {TypeNames, getTypeByScalarName} from '../types';
 import {DirectiveName} from './constant';
 import {buildSchemaFromFile} from './schema';
 import {
@@ -80,20 +82,28 @@ export function getAllEntitiesRelations(_schema: GraphQLSchema | string | null):
   const modelRelations = {models: [], relations: [], enums: [...enums.values()]} as GraphQLModelsRelationsEnums;
   const derivedFrom = schema.getDirective('derivedFrom');
   const indexDirective = schema.getDirective('index');
+  assert(derivedFrom && indexDirective, 'derivedFrom and index directives are required');
   for (const entity of entities) {
     const newModel: GraphQLModelsType = {
       name: entity.name,
-      description: entity.description,
+      description: entity.description ?? undefined,
       fields: [],
       indexes: [],
     };
     const entityFields = Object.values(entity.getFields());
 
+    const idField = entityFields.find((field) => field.name === 'id');
+    if (!idField) {
+      throw new Error(`Entity "${entity.name}" is missing required id field.`);
+    } else if (idField.type.toString() !== 'ID!') {
+      throw new Error(`Entity "${entity.name}" type must be ID, received ${idField.type.toString()}`);
+    }
+
     const fkNameSet: string[] = [];
     for (const field of entityFields) {
       const typeString = extractType(field.type);
-      const derivedFromDirectValues = getDirectiveValues(derivedFrom, field.astNode);
-      const indexDirectiveVal = getDirectiveValues(indexDirective, field.astNode);
+      const derivedFromDirectValues = field.astNode ? getDirectiveValues(derivedFrom, field.astNode) : undefined;
+      const indexDirectiveVal = field.astNode ? getDirectiveValues(indexDirective, field.astNode) : undefined;
 
       //If is a basic scalar type
       const typeClass = getTypeByScalarName(typeString);
@@ -104,7 +114,7 @@ export function getAllEntitiesRelations(_schema: GraphQLSchema | string | null):
       else if (enums.has(typeString)) {
         newModel.fields.push({
           type: typeString,
-          description: field.description,
+          description: field.description ?? undefined,
           isEnum: true,
           isArray: isListType(isNonNullType(field.type) ? getNullableType(field.type) : field.type),
           nullable: !isNonNullType(field.type),
@@ -120,6 +130,32 @@ export function getAllEntitiesRelations(_schema: GraphQLSchema | string | null):
           to: typeString,
           foreignKey: `${field.name}Id`,
         } as GraphQLRelationsType);
+
+        // Ensures that foreign keys are linked correctly
+        if (
+          field.astNode &&
+          (field.astNode.type.kind === 'ListType' ||
+            (field.astNode.type.kind === 'NonNullType' && field.astNode.type.type.kind === 'ListType'))
+        ) {
+          const resolveName = (type: TypeNode): string => {
+            switch (type.kind) {
+              case 'NamedType':
+                return type.name.value;
+              case 'NonNullType':
+                return resolveName(type.type);
+              case 'ListType':
+                return resolveName(type.type);
+              default:
+                // Any case in case future adds new kind
+                throw new Error(`Unandled node kind: ${(type as any).kind}`);
+            }
+          };
+
+          throw new Error(
+            `Field "${field.name}" on entity "${newModel.name}" is missing "derivedFrom" directive. Please also make sure "${resolveName(field.astNode.type)}" has a field of type "${newModel.name}".`
+          );
+        }
+
         newModel.indexes.push({
           unique: false,
           fields: [`${field.name}Id`],
@@ -140,10 +176,11 @@ export function getAllEntitiesRelations(_schema: GraphQLSchema | string | null):
       // If is jsonField
       else if (jsonObjects.map((json) => json.name).includes(typeString)) {
         const jsonObject = jsonObjects.find((object) => object.name === typeString);
+        assert(jsonObject, `Json object ${typeString} not found`);
         const jsonObjectType = setJsonObjectType(jsonObject, jsonObjects);
         newModel.fields.push(packJSONField(typeString, field, jsonObjectType));
 
-        const directive = jsonObject.astNode.directives.find(({name: {value}}) => value === DirectiveName.JsonField);
+        const directive = jsonObject.astNode?.directives?.find(({name: {value}}) => value === DirectiveName.JsonField);
         const argValue = directive?.arguments?.find((arg) => arg.name.value === 'indexed')?.value;
         // For backwards compatibility if the argument is not defined then the index will be added
         if (!argValue || (isBooleanValueNode(argValue) && argValue.value !== false)) {
@@ -180,10 +217,12 @@ export function getAllEntitiesRelations(_schema: GraphQLSchema | string | null):
 
     // Composite Indexes
     const compositeIndexDirective = schema.getDirective('compositeIndexes');
+    assert(compositeIndexDirective, 'compositeIndexes directive is required');
+    assert(entity.astNode, 'Entity astNode is required');
     const compositeIndexDirectiveVal = getDirectiveValues(compositeIndexDirective, entity.astNode) as {
       fields?: string[][];
     };
-    if (compositeIndexDirectiveVal?.fields.length) {
+    if (compositeIndexDirectiveVal?.fields && compositeIndexDirectiveVal?.fields.length) {
       const duplicateIndexes = findDuplicateStringArray(compositeIndexDirectiveVal.fields);
       if (duplicateIndexes.length) {
         throw new Error(
@@ -200,6 +239,7 @@ export function getAllEntitiesRelations(_schema: GraphQLSchema | string | null):
 
     // Fulltext Search
     const fullTextDirective = schema.getDirective('fullText');
+    assert(fullTextDirective, 'fullText directive is required');
     const fullTextDirectiveVal = getDirectiveValues(fullTextDirective, entity.astNode) as GraphQLFullTextType;
 
     if (fullTextDirectiveVal) {
@@ -276,7 +316,7 @@ function packEntityField(
   return {
     name: isForeignKey ? `${field.name}Id` : field.name,
     type: isForeignKey ? FieldScalar.String : typeString,
-    description: field.description,
+    description: field.description ?? undefined,
     isArray: isListType(isNonNullType(field.type) ? getNullableType(field.type) : field.type),
     nullable: !isNonNullType(field.type),
     isEnum: false,
@@ -291,7 +331,7 @@ function packJSONField(
   return {
     name: field.name,
     type: 'Json',
-    description: field.description,
+    description: field.description ?? undefined,
     jsonInterface: jsonObject,
     isArray: isListType(isNonNullType(field.type) ? getNullableType(field.type) : field.type),
     nullable: !isNonNullType(field.type),
@@ -310,16 +350,12 @@ export function setJsonObjectType(
   for (const field of Object.values(jsonObject.getFields())) {
     //check if field is also json
     const typeString = extractType(field.type);
-    const isJsonType = jsonObjects.map((json) => json.name).includes(typeString);
+    const jsonType = jsonObjects.find((json) => json.name === typeString);
+
     graphQLJsonObject.fields.push({
       name: field.name,
-      type: isJsonType ? 'Json' : extractType(field.type),
-      jsonInterface: isJsonType
-        ? setJsonObjectType(
-            jsonObjects.find((object) => object.name === typeString),
-            jsonObjects
-          )
-        : undefined,
+      type: jsonType ? 'Json' : extractType(field.type),
+      jsonInterface: jsonType ? setJsonObjectType(jsonType, jsonObjects) : undefined,
       nullable: !isNonNullType(field.type),
       isArray: isListType(isNonNullType(field.type) ? getNullableType(field.type) : field.type),
     } as GraphQLJsonFieldType);