+++ packages/cli/CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.14.0] - 2024-06-18
 ### Removed
 - Deprecated flare SDK (#2428)
 
@@ -609,7 +610,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - support subcommand codegen
 - support subcommand init
 
-[Unreleased]: https://github.com/subquery/subql/compare/cli/4.13.1...HEAD
+[Unreleased]: https://github.com/subquery/subql/compare/cli/4.14.0...HEAD
+[4.14.0]: https://github.com/subquery/subql/compare/cli/4.13.1...cli/4.14.0
 [4.13.1]: https://github.com/subquery/subql/compare/cli/4.13.0...cli/4.13.1
 [4.13.0]: https://github.com/subquery/subql/compare/cli/4.12.0...cli/4.13.0
 [4.12.0]: https://github.com/subquery/subql/compare/cli/4.11.0...cli/4.12.0