+++ packages/utils/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Better error messages for invalid graphql schemas (#2458)
+
+### Added
+- enable ts strict mode
 
 ## [2.11.0] - 2024-06-12
 ### Changed